backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.20
pydantic>=2.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.6
stripe>=10.6.0
psycopg2-binary>=2.9.9
PyJWT>=2.9.0
pytest>=8.3.2
pytest-asyncio>=1.2.0
pytest-cov>=4.1.0
alembic>=1.13.0
redis>=5.0.0
passlib[bcrypt]>=1.7.4
Pillow>=10.0.0
aiofiles>=23.0.0
psutil>=5.9.0
email-validator>=2.0.0
authlib>=1.3.1
httpx>=0.27.0

# Security updates
jinja2>=3.1.4
aiohttp>=3.9.5
idna>=3.7
urllib3>=2.2.2
certifi>=2024.7.4
requests>=2.32.2
setuptools>=70.0.0
cryptography>=42.0.8
